Fix the specification of F.min and F.max
The correct implementation then required to add is_nan.

The specification of F.neg has been strengthened since it was required
in Tactic.v (Lemma contains_bound_ar that uses meet and F.neg in
combination) to know that F.neg of a nan is a nan.

A few Admitted added in Interval/Float.v but these proof will be
broken again when improving F.cmp and I don't want to do the job
twice.