Fix and prove I.midpoint
I.midpoint [x, +oo] was defined as 2x when x > 0. I had to replace
this with x because 2x could overflow and we want midpoint to be real.

A more elaborate solution would be to return min x (mul_UP 2 x) but
this would requir a much more precise specification of mul_UP
(currently mul_UP 2 x) can perfectly return a NaN, even for x real.